refactor grpc: refactor grpc logging commit_hash:736dbd4ec685f07d363e4d2c77fbf0c00db9edc4
