+++ .mapping.json
@@ -1972,6 +1972,8 @@
   "grpc/src/ugrpc/impl/code_statistics.cpp":"taxi/uservices/userver/grpc/src/ugrpc/impl/code_statistics.cpp",
   "grpc/src/ugrpc/impl/completion_queue_pool_base.cpp":"taxi/uservices/userver/grpc/src/ugrpc/impl/completion_queue_pool_base.cpp",
   "grpc/src/ugrpc/impl/deadline_timepoint.cpp":"taxi/uservices/userver/grpc/src/ugrpc/impl/deadline_timepoint.cpp",
+  "grpc/src/ugrpc/impl/grpc_native_logging.cpp":"taxi/uservices/userver/grpc/src/ugrpc/impl/grpc_native_logging.cpp",
+  "grpc/src/ugrpc/impl/grpc_native_logging.hpp":"taxi/uservices/userver/grpc/src/ugrpc/impl/grpc_native_logging.hpp",
   "grpc/src/ugrpc/impl/grpc_string_logging.hpp":"taxi/uservices/userver/grpc/src/ugrpc/impl/grpc_string_logging.hpp",
   "grpc/src/ugrpc/impl/internal_tag.hpp":"taxi/uservices/userver/grpc/src/ugrpc/impl/internal_tag.hpp",
   "grpc/src/ugrpc/impl/logging.cpp":"taxi/uservices/userver/grpc/src/ugrpc/impl/logging.cpp",
