+++ grpc/src/ugrpc/client/common_component.cpp
@@ -5,7 +5,7 @@
 #include <userver/logging/level_serialization.hpp>
 #include <userver/yaml_config/merge_schemas.hpp>
 
-#include <ugrpc/impl/logging.hpp>
+#include <ugrpc/impl/grpc_native_logging.hpp>
 #include <userver/ugrpc/client/middlewares/base.hpp>
 #include <userver/ugrpc/server/server_component.hpp>
