+++ grpc/src/ugrpc/client/middlewares/log/component.cpp
@@ -11,9 +11,9 @@ namespace ugrpc::client::middlewares::log {
 
 Settings Parse(const yaml_config::YamlConfig& config, formats::parse::To<Settings>) {
     Settings settings;
-    settings.max_msg_size = config["msg-size-log-limit"].As<std::size_t>(settings.max_msg_size);
     settings.log_level = config["log-level"].As<logging::Level>(settings.log_level);
-    settings.log_level = config["msg-log-level"].As<logging::Level>(settings.msg_log_level);
+    settings.msg_log_level = config["msg-log-level"].As<logging::Level>(settings.msg_log_level);
+    settings.max_msg_size = config["msg-size-log-limit"].As<std::size_t>(settings.max_msg_size);
     settings.trim_secrets = config["trim-secrets"].As<bool>(settings.trim_secrets);
     return settings;
 }
