+++ grpc/src/ugrpc/impl/protobuf_utils.hpp
@@ -1,40 +1,23 @@
 #pragma once
 
-#include <cstddef>
-#include <string>
+#include <google/protobuf/descriptor.h>
+#include <google/protobuf/message.h>
 
-namespace google::protobuf {
-
-class Message;
-class FieldDescriptor;
-
-}  // namespace google::protobuf
+#include <userver/field_options.pb.h>
 
 USERVER_NAMESPACE_BEGIN
 
 namespace ugrpc::impl {
 
+const userver::FieldOptions& GetFieldOptions(const google::protobuf::FieldDescriptor& field);
+
 bool IsMessage(const google::protobuf::FieldDescriptor& field);
+
+bool HasSecrets(const google::protobuf::Message& message);
 
 /// @warning This causes a segmentation fault for messages containing optional fields in protobuf versions prior to 3.13
 /// See https://github.com/protocolbuffers/protobuf/issues/7801
 void TrimSecrets(google::protobuf::Message& message);
-
-std::string ToString(const google::protobuf::Message& message, std::size_t max_msg_size);
-
-std::string ToJsonString(const google::protobuf::Message& message, std::size_t max_msg_size);
-
-// Same as ToString but also trims the secrets
-///
-/// @warning This causes a segmentation fault for messages containing optional fields in protobuf versions prior to 3.13
-/// See https://github.com/protocolbuffers/protobuf/issues/7801
-std::string ToLogString(const google::protobuf::Message& message, std::size_t max_msg_size);
-
-// Same as ToJsonString but also trims the secrets
-///
-/// @warning This causes a segmentation fault for messages containing optional fields in protobuf versions prior to 3.13
-/// See https://github.com/protocolbuffers/protobuf/issues/7801
-std::string ToJsonLogString(const google::protobuf::Message& message, std::size_t max_msg_size);
 
 }  // namespace ugrpc::impl
