+++ grpc/src/ugrpc/server/middlewares/log/middleware.cpp
@@ -5,7 +5,7 @@
 #include <userver/tracing/span.hpp>
 #include <userver/yaml_config/yaml_config.hpp>
 
-#include <ugrpc/impl/protobuf_utils.hpp>
+#include <ugrpc/impl/logging.hpp>
 
 USERVER_NAMESPACE_BEGIN
 
@@ -22,27 +22,13 @@ bool IsResponseStream(CallKind kind) {
 }
 
 std::string GetMessageForLogging(const google::protobuf::Message& message, const Settings& settings) {
-    if (logging::ShouldLog(settings.msg_log_level)) {
-        if (settings.trim_secrets) {
-            return ugrpc::impl::ToJsonLogString(message, settings.max_msg_size);
-        } else {
-            return ugrpc::impl::ToJsonString(message, settings.max_msg_size);
-        }
-    } else {
-        return "hidden by log level";
-    }
+    return ugrpc::impl::GetMessageForLogging(
+        message,
+        ugrpc::impl::MessageLoggingOptions{settings.msg_log_level, settings.max_msg_size, settings.trim_secrets}
+    );
 }
 
 }  // namespace
-
-Settings Parse(const yaml_config::YamlConfig& config, formats::parse::To<Settings>) {
-    Settings settings;
-    settings.max_msg_size = config["msg-size-log-limit"].As<std::size_t>(settings.max_msg_size);
-    settings.msg_log_level = config["msg-log-level"].As<logging::Level>(settings.msg_log_level);
-    settings.local_log_level = config["log-level"].As<std::optional<logging::Level>>();
-    settings.trim_secrets = config["trim-secrets"].As<bool>(settings.trim_secrets);
-    return settings;
-}
 
 Middleware::Middleware(const Settings& settings) : settings_(settings) {}
