+++ grpc/src/ugrpc/server/server.cpp
@@ -17,7 +17,7 @@
 #include <userver/utils/assert.hpp>
 #include <userver/utils/fixed_array.hpp>
 
-#include <ugrpc/impl/logging.hpp>
+#include <ugrpc/impl/grpc_native_logging.hpp>
 #include <ugrpc/server/impl/generic_service_worker.hpp>
 #include <ugrpc/server/impl/parse_config.hpp>
 #include <userver/ugrpc/impl/deadline_timepoint.hpp>
